Deleted folders and Files, Modified Cards, Organize folders and rename somes id and class - Deleted ADD Card folder and components files Status.jsx and New.jsx - Modified arquiteture of Cards (HTML) - Add folders Hooks and pages - rename somes classes and ids
