+++ client/src/App.tsx
@@ -1,16 +1,11 @@
 import './css/App.css'
-import Dashboard from './components/Dashboard'
-import AddCard from './addCard/Dashboard'
+import Dashboard from './pages/Dashboard'
 import {getCard} from './service/api'
-
-
 
 export default function App() {
 	return(
 		<>
 			<Dashboard/>
-			
-			<AddCard/>
 		</>
 	)
 }
