+++ client/src/components/Cards.tsx
@@ -25,32 +25,14 @@ export default function Cards({dataCards, buttonIndex}: Props) {
             setResponse("Não há mais cards.")
         }
     }, [buttonIndex])
-    
-    function handleClick() {
-        document.getElementById("dashboard__content__SeeResponse")!.style.display = "none"
-        document.getElementById("dashboard__content__response")!.style.display = "inline"
-     
-    }
     return(
-        <>
-        <section className="dashboard__content__cards">
-            
-            <section className="dashboard__content__questFlex" id="dashboard__content__quest"> 
-                <div id="dashboard__content__quest_output">
-                    <p className="dashboard__content__flexText">{question}</p>
-                </div>
-            </section> 
-            
-            <section className="dashboard__content__questFlex" id="dashboard__content__section_response">
-                <div id="dashboard__content__response_output">
-                    <p id="dashboard__content__response" style={{display: "none"}}>{response}</p>
-                    <div className="dashboard__content__see">                    
-                        <button className="buttons" id="dashboard__content__SeeResponse" onClick={handleClick}>Resposta</button>    
-                    </div>
-                </div>
-            </section>
-            
-        </section>
-        </>
+        <div className="card">
+            <h1 className="card_title">Title</h1>
+            <p className="card_quest">{question}</p>
+            <div className="card_footer">
+                <span>Total: <a></a></span> 
+                <div> Virar Card</div>
+            </div>
+        </div>
     )
 }
